Implement barcode event tracking and refactor event handling logic
